SQL Create obj Script

+++ ProjectFiles/SQL Files dbo/BikeStores Sample Database - create objects.sql
@@ -27,7 +27,7 @@ CREATE TABLE products (
 	brand_id INT NOT NULL,
 	category_id INT NOT NULL,
 	model_year SMALLINT NOT NULL,
-	list_price DECIMAL (10, 2) NOT NULL,
+	list_price FLOAT NOT NULL,
 	FOREIGN KEY (category_id) REFERENCES  categories (category_id) ON DELETE CASCADE ON UPDATE CASCADE,
 	FOREIGN KEY (brand_id) REFERENCES  brands (brand_id) ON DELETE CASCADE ON UPDATE CASCADE
 );
@@ -88,8 +88,8 @@ CREATE TABLE order_items (
 	item_id INT,
 	product_id INT NOT NULL,
 	quantity INT NOT NULL,
-	list_price DECIMAL (10, 2) NOT NULL,
-	discount DECIMAL (4, 2) NOT NULL DEFAULT 0,
+	list_price FLOAT NOT NULL,
+	discount FLOAT NOT NULL DEFAULT 0,
 	PRIMARY KEY (order_id, item_id),
 	FOREIGN KEY (order_id) REFERENCES  orders (order_id) ON DELETE CASCADE ON UPDATE CASCADE,
 	FOREIGN KEY (product_id) REFERENCES  products (product_id) ON DELETE CASCADE ON UPDATE CASCADE
